third_party/blink/web_tests/svg/custom/tearoffs-with-tearoffs.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests whether SVG tearoff objects with tearoffs update properly.");

var svgDoc = document.implementation.createDocument("http://www.w3.org/2000/svg", "svg", null);
var transform = svgDoc.documentElement.createSVGTransform();

shouldBe("transform.matrix.a", "1");
transform.matrix.a = 2;
shouldBe("transform.matrix.a", "2");

var successfullyParsed = true;
</script>
</body>
</html>
