third_party/blink/web_tests/platform/generic/svg/custom/tearoffs-with-tearoffs-expected.txt

Tests whether SVG tearoff objects with tearoffs update properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS transform.matrix.a is 1
PASS transform.matrix.a is 2
PASS successfullyParsed is true

TEST COMPLETE
